# src/types/genre.ts
export interface Genre {
    /**
     * @isString Only accept string values
     * @minLength 1 Minimum length of 1
     */
    id: string
    /**
     * @isString Only accept string values
     * @minLength 1 Minimum length of 1
     */
    name: string
}